refactor: Pull out common base from Finding and FindingOne
