+++ lib/plugin/components/state-resources/finding/index.js
@@ -1,37 +1,9 @@
-'use strict'
+const FindStateResourceBase = require('./find-base')
 
-const boom = require('@hapi/boom')
-const debug = require('debug')('findingOne')
-
-module.exports = class Finding {
-  init (resourceConfig, env) {
-    this.modelId = resourceConfig.modelId
-    this.filterTemplate = resourceConfig.filter || {}
-    const models = env.bootedServices.storage.models
-    if (Object.prototype.hasOwnProperty.call(models, this.modelId)) {
-      this.model = models[this.modelId]
-    } else {
-      throw boom.notFound('Unable to initialize Finding state... unknown model \'' + this.modelId + '\'', { modelId: this.modelId })
-    }
-  }
-
-  run (event, context) {
-    const filter = context.resolveInputPaths(event, this.filterTemplate)
-    debug(`Filtering model '${this.modelId}' ${JSON.stringify(filter)} - (executionName='${context.executionName}')`)
-    this.model.find(
-      filter,
-      function (err, doc) {
-        if (err) {
-          context.sendTaskFailure(
-            {
-              error: 'FAILED_TO_FIND',
-              cause: err
-            }
-          )
-        } else {
-          context.sendTaskSuccess(doc)
-        }
-      }
-    )
+class Finding extends FindStateResourceBase {
+  constructor () {
+    super('Finding', 'find')
   }
 }
+
+module.exports = Finding
